requirements-dev.txt
requests>=2.32.0
python-dotenv>=1.0.1
ollama>=0.3.1
